Vastly Increased Gallery Performance
Gallery page now uses a small thumbnail picture for displaying in the grid, and the modal that opens when an image is pressed uses a larger but still processed picture that loads on press. There is now also a download button in the modal screen to download the picture, the downloaded picture is the unprocessed original photo.

Changed 3 scripts to make this work
1. htmlUpdater.py now uses the .webp extension and ensures no larger images get caught
2. PicConvert.py now creates 2 .webp pics, a thumbnail and a large preview for picture modal
3. The modal script now includes download button functionality and points to the correct pictures for modal and download.